fix: solved -werror -warning -wall errors

+++ src/ServerEngine/Location.cpp
@@ -51,9 +51,9 @@ int Location::setErrorPages(std::list<Node>::iterator &it) {
 		return 1;
 	}
 
-	for (int i = 0; i < stash.size() - 1; i++) {
+	for (size_t i = 0; i < stash.size() - 1; i++) {
 		// Check all digits
-		for (int j = 0; j < stash[i].length(); j++) {
+		for (size_t j = 0; j < stash[i].length(); j++) {
 			if (!isdigit(stash[i][j])) {
 				log(std::cerr, MsgType::ERROR, "Invalid error code", stash[i]);
 				return 1;
@@ -105,7 +105,7 @@ int Location::setClientMaxBodySize(std::list<Node>::iterator &it) {
 	int flag = NUMBER;
 	std::string body_size = stash.back();
 	stash.clear();
-	for (int i = 0; i < body_size.length(); i++) {
+	for (size_t i = 0; i < body_size.length(); i++) {
 		if (body_size[i] == 'k' || body_size[i] == 'K' || \
 			body_size[i] == 'm' || body_size[i] == 'M' || \
 			body_size[i] == 'g' || body_size[i] == 'G') {
